+++ src/ByteDev.Common.UnitTests/Reflection/ReflectionAttributeExtensionsTest.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Reflection;
 using ByteDev.Common.Reflection;
 using NUnit.Framework;
 
@@ -8,34 +9,90 @@ namespace ByteDev.Common.UnitTests.Reflection
     public class ReflectionAttributeExtensionsTest
     {
         [TestFixture]
-        public class HasAttribute
+        public class HasAttribute_Type : ReflectionAttributeExtensionsTest
         {
             [Test]
-            public void WhenClassMethodHasAttribute_ThenReturnTrue()
+            public void WhenTypeIsNull_ThenThrowException()
             {
-                var methodUnderTest = typeof(DummyWithMethods).GetMethod("MethodWithAttribute");
+                Assert.Throws<ArgumentNullException>(() => (null as Type).HasAttribute<UsedAttribute>());
+            }
 
-                var result = methodUnderTest.HasAttribute<UsedAttribute>();
+            [Test]
+            public void WhenTypeHasAttribute_ThenReturnTrue()
+            {
+                Type sut = typeof(DummyWithClassAttribute);
+
+                var result = sut.HasAttribute<UsedAttribute>();
 
                 Assert.That(result, Is.True);
             }
 
             [Test]
-            public void WhenClassMethodDoesntHaveAttribute_ThenReturnFalse()
+            public void WhenTypeDoesNotHaveAttribute_ThenReturnFalse()
             {
-                var methodUnderTest = typeof(DummyWithMethods).GetMethod("MethodWithNoAttribute");
+                Type sut = typeof(DummyWithClassAttribute);
 
-                var result = methodUnderTest.HasAttribute<UsedAttribute>();
+                var result = sut.HasAttribute<NotUsedAttribute>();
 
                 Assert.That(result, Is.False);
             }
 
             [Test]
+            public void WhenTypesMethodsHasAttribute_ThenReturnFalse()
+            {
+                Type sut = typeof(DummyWithMethods);
+
+                var result = sut.HasAttribute<UsedAttribute>();
+
+                Assert.That(result, Is.False);
+            }
+        }
+
+        [TestFixture]
+        public class HasAttribute_MemberInfo : ReflectionAttributeExtensionsTest
+        {
+            [Test]
+            public void WhenTypeIsNull_ThenThrowException()
+            {
+                Assert.Throws<ArgumentNullException>(() => (null as MemberInfo).HasAttribute<UsedAttribute>());
+            }
+
+            [Test]
+            public void WhenMethodHasAttribute_ThenReturnTrue()
+            {
+                MethodInfo sut = typeof(DummyWithMethods).GetMethod("MethodWithAttribute");
+
+                var result = sut.HasAttribute<UsedAttribute>();
+
+                Assert.That(result, Is.True);
+            }
+
+            [Test]
+            public void WhenMethodDoesNotHaveAttribute_ThenReturnFalse()
+            {
+                MethodInfo sut = typeof(DummyWithMethods).GetMethod("MethodWithNoAttribute");
+
+                var result = sut.HasAttribute<UsedAttribute>();
+
+                Assert.That(result, Is.False);
+            }
+        }
+
+        [TestFixture]
+        public class HasAttribute_Object : ReflectionAttributeExtensionsTest
+        {
+            [Test]
+            public void WhenTypeIsNull_ThenThrowException()
+            {
+                Assert.Throws<ArgumentNullException>(() => (null as object).HasAttribute<UsedAttribute>());
+            }
+
+            [Test]
             public void WhenObjectHasAttribute_ThenReturnTrue()
             {
-                var typeUnderTest = new DummyWithClassAttribute();
+                object sut = new DummyWithClassAttribute();
 
-                var result = typeUnderTest.HasAttribute<UsedAttribute>();
+                var result = sut.HasAttribute<UsedAttribute>();
 
                 Assert.That(result, Is.True);
             }
@@ -43,15 +100,19 @@ namespace ByteDev.Common.UnitTests.Reflection
             [Test]
             public void WhenObjectHasNoAttribute_ThenReturnFalse()
             {
-                var typeUnderTest = new DummyWithMethods();
+                object sut = new DummyWithMethods();
 
-                var result = typeUnderTest.HasAttribute<UsedAttribute>();
+                var result = sut.HasAttribute<UsedAttribute>();
 
                 Assert.That(result, Is.False);
             }
         }
         
         public class UsedAttribute : Attribute
+        {
+        }
+
+        public class NotUsedAttribute : Attribute
         {
         }